business/requirements.txt
fastapi[all]
sqlalchemy[asyncio]
asyncpg
httpx
pydantic
pydantic-settings
python-dateutil
aio-pika
numpy
pandas